requirements.txt
python-dotenv
tensorflow==2.1.0
numpy
sklearn
matplotlib
pandas
opencv-python
face_recognition